Accepted patch #3303614 from McSpider, adding OpenFFA, fixed robot targeting of robots, bumped PROTO->205
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21542 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/playing.cxx
@@ -740,7 +740,7 @@ static void doKeyPlaying(const BzfKeyEvent& key, bool pressed, bool haveBinding)
       char name[32];
       int msgno = (key.button - BzfKeyEvent::F1) + 1;
       void* buf = messageMessage;
-      if (key.shift == BzfKeyEvent::ControlKey) {
+      if (key.shift == BzfKeyEvent::ControlKey && world->allowTeams()) {
 	sprintf(name, "quickTeamMessage%d", msgno);
 	buf = nboPackUByte(buf, TeamToPlayerId(myTank->getTeam()));
       } else {
@@ -2185,8 +2185,8 @@ static void		handleServerMessage(bool human, uint16_t code,
 	}
 	if (victimPlayer && killerLocal != victimPlayer) {
 	  if ((victimPlayer->getTeam() == killerLocal->getTeam()) &&
-	      (killerLocal->getTeam() != RogueTeam) &&
-              !(killerPlayer == myTank && wasRabbit)) {
+        (killerLocal->getTeam() != RogueTeam) && !(killerPlayer == myTank && wasRabbit)
+        && World::getWorld()->allowTeams()) {
             // teamkill
 	    if (killerPlayer == myTank) {
 	      hud->setAlert(1, "Don't kill teammates!!!", 3.0f, true);
@@ -2258,7 +2258,7 @@ static void		handleServerMessage(bool human, uint16_t code,
 	}
 	else {
 	  std::string playerStr;
-	  if ((killerPlayer->getTeam() == victimPlayer->getTeam()) &&
+	  if (World::getWorld()->allowTeams() && (killerPlayer->getTeam() == victimPlayer->getTeam()) &&
 	      (killerPlayer->getTeam() != RogueTeam) &&
 	      (killerPlayer->getTeam() != ObserverTeam)) {
 	    playerStr += "teammate ";
@@ -2296,28 +2296,31 @@ static void		handleServerMessage(bool human, uint16_t code,
 	}
       }
 
-      // blow up if killer has genocide flag and i'm on same team as victim
-      // (and we're not rogues, unless in rabbit mode)
-      if (human && killerPlayer && victimPlayer && victimPlayer != myTank &&
-	  (victimPlayer->getTeam() == myTank->getTeam()) &&
-	  (myTank->getTeam() != RogueTeam) && shotId >= 0) {
-	// now see if shot was fired with a GenocideFlag
-	const ShotPath* shot = killerPlayer->getShot(int(shotId));
-	if (shot && shot->getFlag() == Flags::Genocide) {
-	  gotBlowedUp(myTank, GenocideEffect, killerPlayer->getId());
-	}
-      }
-
+      if (World::getWorld()->allowTeams())  // geno only works in team games :)
+      {
+        // blow up if killer has genocide flag and i'm on same team as victim
+        // (and we're not rogues, unless in rabbit mode)
+        if (human && killerPlayer && victimPlayer && victimPlayer != myTank &&
+            (victimPlayer->getTeam() == myTank->getTeam()) &&
+            (myTank->getTeam() != RogueTeam) && shotId >= 0) {
+          // now see if shot was fired with a GenocideFlag
+          const ShotPath* shot = killerPlayer->getShot(int(shotId));
+          if (shot && shot->getFlag() == Flags::Genocide) {
+            gotBlowedUp(myTank, GenocideEffect, killerPlayer->getId());
+          }
+        }
+      
 #ifdef ROBOT
-      // blow up robots on victim's team if shot was genocide
-      if (killerPlayer && victimPlayer && shotId >= 0) {
-	const ShotPath* shot = killerPlayer->getShot(int(shotId));
-	if (shot && shot->getFlag() == Flags::Genocide)
-	  for (int i = 0; i < numRobots; i++)
-	    if (robots[i] && victimPlayer != robots[i] &&
-		victimPlayer->getTeam() == robots[i]->getTeam() &&
-		robots[i]->getTeam() != RogueTeam)
-	      gotBlowedUp(robots[i], GenocideEffect, killerPlayer->getId());
+        // blow up robots on victim's team if shot was genocide
+        if (killerPlayer && victimPlayer && shotId >= 0) {
+          const ShotPath* shot = killerPlayer->getShot(int(shotId));
+          if (shot && shot->getFlag() == Flags::Genocide)
+            for (int i = 0; i < numRobots; i++)
+            if (robots[i] && victimPlayer != robots[i] &&
+                victimPlayer->getTeam() == robots[i]->getTeam() &&
+                robots[i]->getTeam() != RogueTeam)
+              gotBlowedUp(robots[i], GenocideEffect, killerPlayer->getId());
+        }
       }
 #endif
 
@@ -3487,8 +3490,7 @@ static bool		gotBlowedUp(BaseLocalPlayer* tank,
 	  TeamColor team = killerPlayer->getTeam();
 	  if (hit)
 	    team = hit->getTeam();
-	  if ((myTank->getTeam() == team) &&
-	      (team != RogueTeam) && (team != ObserverTeam)) {
+	  if (World::getWorld()->allowTeams() && (myTank->getTeam() == team) && (team != RogueTeam) && (team != ObserverTeam)) {
 	    blowedUpNotice += "teammate " ;
 	    blowedUpNotice += killerPlayer->getCallSign();
 	  } else {
@@ -4025,8 +4027,7 @@ static void		setRobotTarget(RobotPlayer* robot)
   float bestPriority = 0.0f;
   for (int j = 0; j < curMaxPlayers; j++)
     if (player[j] && player[j]->getId() != robot->getId() &&
-	player[j]->isAlive() &&
-	((robot->getTeam() == RogueTeam) || player[j]->getTeam() != robot->getTeam())) {
+	player[j]->isAlive() && robot->validTeamTarget(player[j])) {
 
       if (player[j]->isPhantomZoned() && !robot->isPhantomZoned())
 	continue;
@@ -4047,7 +4048,7 @@ static void		setRobotTarget(RobotPlayer* robot)
       }
     }
   if (myTank->isAlive() &&
-      ((robot->getTeam() == RogueTeam) || myTank->getTeam() != robot->getTeam())) {
+      ((robot->getTeam() == RogueTeam) ||  robot->validTeamTarget(myTank))) {
     const float priority = robot->getTargetPriority(myTank);
     if (priority > bestPriority) {
       bestTarget = myTank;
